Move `cache.ts` to `cli/output-cache.ts`
This change should hopefully make the purpose of this module clear.
