+++ src/testing-utils.ts
@@ -14,8 +14,8 @@ import { ActionsEnv, getActionVersion } from "./actions-util";
 import { AnalysisKind } from "./analyses";
 import * as apiClient from "./api-client";
 import { GitHubApiDetails } from "./api-client";
-import { resetCachedCommandOutputs } from "./cache";
 import { CachingKind } from "./caching-utils";
+import { resetCachedCommandOutputs } from "./cli/output-cache";
 import * as codeql from "./codeql";
 import { Config } from "./config-utils";
 import * as defaults from "./defaults.json";
